Added tests for fim_diff_process_delete to files, registry keys and values

+++ src/unit_tests/wrappers/libc/stdio_wrappers.c
@@ -107,7 +107,7 @@ extern size_t __real_fread(void *ptr, size_t size, size_t n, FILE *stream);
 size_t __wrap_fread(void *ptr, size_t size, size_t n, FILE *stream) {
     if (test_mode) {
         strncpy((char *) ptr, mock_type(char *), n);
-        int ret = mock();
+        size_t ret = mock();
         if (ret > n){
             return n;
         } else {
